fixing some tedious int to double compilation errors.

+++ include/cubature_1d.h
@@ -40,7 +40,7 @@ public:
   // #### methods ####
 
   // method to get number of cubature_1d points
-  double get_n_pts(void);
+  int get_n_pts(void);
 
   // method to get r location of cubature_1d point
   double get_r(int in_pos);
