+++ src/cubature_tet.cpp
@@ -89,7 +89,7 @@ void cubature_tet::set_rule(int in_rule)
 
 // method to get number of cubature points
 
-double cubature_tet::get_n_pts(void)
+int cubature_tet::get_n_pts(void)
 {
   return n_pts;
 }
